subvenciones-vivienda/scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
supabase==2.3.0
python-dotenv==1.0.0
schedule==1.2.1
feedparser==6.0.10
resend==0.7.0
